PDF-to-Audio/requirements.txt
pypdf>=3.0.0
pdfplumber>=0.9.0
pyttsx3>=2.90
gTTS>=2.3.0
pydub>=0.25.1
tqdm>=4.65.0
Pillow>=9.5.0
pytesseract>=0.3.10
pdf2image>=1.16.0
edge-tts>=6.1.0
